requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.39.0
sqlmodel==0.0.31
pillow==11.3.0
python-multipart==0.0.20
requests==2.32.4
python-dotenv==1.2.1
better_profanity==0.7.0
pydantic==2.12.5